Рефакторинг часть 1 Done
